requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-multipart==0.0.18

# Telegram bot
python-telegram-bot==20.7

# LLM & Agent
anthropic==0.39.0
langgraph==0.2.60
langchain-anthropic==0.3.1
langchain-core==0.3.28

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Validation
pydantic==2.10.3
pydantic-settings==2.7.0

# Logging
loguru==0.7.3

# Utils
python-dotenv==1.0.1
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# LangSmith (optional)
langsmith==0.2.3
